+++ src/main/java/xyz/tomorrowlearncamp/bookking/domain/user/controller/UserController.java
@@ -1,0 +1,24 @@
+package xyz.tomorrowlearncamp.bookking.domain.user.controller;
+
+import lombok.RequiredArgsConstructor;
+import org.springframework.http.ResponseEntity;
+import org.springframework.security.core.annotation.AuthenticationPrincipal;
+import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RestController;
+import xyz.tomorrowlearncamp.bookking.domain.user.auth.dto.AuthUser;
+import xyz.tomorrowlearncamp.bookking.domain.user.dto.UserResponse;
+import xyz.tomorrowlearncamp.bookking.domain.user.service.UserService;
+
+@RestController
+@RequiredArgsConstructor
+@RequestMapping("/api")
+public class UserController {
+    private final UserService userService;
+
+    @GetMapping("/v1/users/myInfo")
+    public ResponseEntity<UserResponse> getMyInfo(@AuthenticationPrincipal AuthUser authUser) {
+        UserResponse response = userService.getMyInfo(authUser.getUserId());
+        return ResponseEntity.ok(response);
+    }
+}
