--- src/main/java/xyz/tomorrowlearncamp/bookking/user/auth/config/JwtAuthenticationFilter.java
@@ -1,78 +1,0 @@
-package xyz.tomorrowlearncamp.bookking.user.auth.config;
-
-import io.jsonwebtoken.Claims;
-import io.jsonwebtoken.ExpiredJwtException;
-import io.jsonwebtoken.MalformedJwtException;
-import io.jsonwebtoken.UnsupportedJwtException;
-import jakarta.servlet.FilterChain;
-import jakarta.servlet.ServletException;
-import jakarta.servlet.http.HttpServletRequest;
-import jakarta.servlet.http.HttpServletResponse;
-import lombok.RequiredArgsConstructor;
-import lombok.extern.slf4j.Slf4j;
-import org.springframework.security.core.context.SecurityContextHolder;
-import org.springframework.stereotype.Component;
-import org.springframework.web.filter.OncePerRequestFilter;
-import xyz.tomorrowlearncamp.bookking.user.auth.dto.AuthUser;
-import xyz.tomorrowlearncamp.bookking.user.enums.UserRole;
-
-import java.io.IOException;
-
-/**
- * 작성자 : 문성준
- * 일시 : 2025.04.03 - v1
- */
-@Slf4j
-@Component
-@RequiredArgsConstructor
-public class JwtAuthenticationFilter extends OncePerRequestFilter {
-
-    private final JwtProvider jwtProvider;
-
-    @Override
-    protected void doFilterInternal(
-            HttpServletRequest request,
-            HttpServletResponse response,
-            FilterChain filterChain
-    ) throws ServletException, IOException {
-        String authHeader = request.getHeader("Authorization");
-
-        if (authHeader != null && authHeader.startsWith("Bearer ")) {
-            String token = jwtProvider.removeBearerPrefix(authHeader);
-            try {
-                Claims claims = jwtProvider.extractClaims(token);
-                if (SecurityContextHolder.getContext().getAuthentication() == null) {
-                    setAuthentication(claims);
-                }
-            } catch (SecurityException | MalformedJwtException e) {
-                log.error("Invalid JWT signature, 유효하지 않는 JWT 서명 입니다.", e);
-                response.sendError(HttpServletResponse.SC_UNAUTHORIZED, "유효하지 않는 JWT 서명입니다.");
-                return;
-            } catch (ExpiredJwtException e) {
-                log.error("Expired JWT token, 만료된 JWT token 입니다.", e);
-                response.sendError(HttpServletResponse.SC_UNAUTHORIZED, "만료된 JWT 토큰입니다.");
-                return;
-            } catch (UnsupportedJwtException e) {
-                log.error("Unsupported JWT token, 지원되지 않는 JWT 토큰 입니다.", e);
-                response.sendError(HttpServletResponse.SC_BAD_REQUEST, "지원되지 않는 JWT 토큰입니다.");
-                return;
-            } catch (Exception e) {
-                log.error("Internal server error", e);
-                response.sendError(HttpServletResponse.SC_INTERNAL_SERVER_ERROR);
-                return;
-            }
-        }
-        filterChain.doFilter(request, response);
-    }
-
-    private void setAuthentication(Claims claims) {
-        Long userId = Long.valueOf(claims.getSubject());
-        String email = claims.get("email", String.class);
-        UserRole userRole = UserRole.of(claims.get("role", String.class));
-
-        AuthUser authUser = AuthUser.of(userId, email, userRole);
-        JwtAuthenticationToken authenticationToken = JwtAuthenticationToken.of(authUser);
-        SecurityContextHolder.getContext().setAuthentication(authenticationToken);
-
-    }
-}
