--- src/main/java/xyz/tomorrowlearncamp/bookking/user/enums/Gender.java
@@ -1,9 +1,0 @@
-package xyz.tomorrowlearncamp.bookking.user.enums;
-
-/**
- * 작성자 : 문성준
- * 일시 : 2025.04.03 - v1
- */
-public enum Gender {
-    MALE, FEMALE
-}
